+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  * remove unused assembler llvm-mc logic to improve performance;
  * add new binary from memory object apis;
  * add the start of text section of relocatable object to function list;
+ * add syscall category in InsnType;
 
 ## v0.1.0
 Initial release.